requirements.txt
nonebot2
nonebot-adapter-cqhttp
httpx
aiohttp
